Rename `returnBook` dir to `returnItem`
